Adiciona tratamento de exceções para usuários e implementa método de deleção no serviço de usuários

+++ backend/src/main/java/com/edvaldo/perdidos_achados/repository/usuario/UsuarioRepositoy.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.edvaldo.perdidos_achados.entity.Usuario;
 
 @Repository
-public interface UsuarioRepositoy  extends JpaRepository<Usuario,Integer>{
+public interface UsuarioRepositoy  extends JpaRepository<Usuario,Long>{
     Optional<Usuario>findByEmail(String email);
 }
